+++ packages/rekapager-api-platform/config/services.php
@@ -11,7 +11,10 @@ declare(strict_types=1);
  * that was distributed with this source code.
  */
 
+use Rekalogika\Rekapager\ApiPlatform\PageNormalizer;
+use Rekalogika\Rekapager\ApiPlatform\PagerFactory;
 use Rekalogika\Rekapager\ApiPlatform\RekapagerOpenApiFactoryDecorator;
+use Rekalogika\Rekapager\Contracts\PageIdentifierEncoderLocatorInterface;
 use Symfony\Component\DependencyInjection\Loader\Configurator\ContainerConfigurator;
 
 use function Symfony\Component\DependencyInjection\Loader\Configurator\service;
@@ -24,5 +27,20 @@ return static function (ContainerConfigurator $containerConfigurator): void {
         ->decorate('api_platform.openapi.factory')
         ->args([
             '$decorated' => service('.inner'),
+        ]);
+
+    $services->set(PagerFactory::class)
+        ->args([
+            '$resourceMetadataFactory' => service('api_platform.metadata.resource.metadata_collection_factory'),
+            '$pageIdentifierEncoderLocator' => service(PageIdentifierEncoderLocatorInterface::class),
+            '$pageParameterName' => '%api_platform.collection.pagination.page_parameter_name%',
+            '$urlGenerationStrategy' => '%api_platform.url_generation_strategy%'
+        ]);
+
+    $services->set('rekalogika.rekapager.api_platform.page_normalizer')
+        ->class(PageNormalizer::class)
+        ->decorate('api_platform.hydra.normalizer.collection')
+        ->args([
+            '$collectionNormalizer' => service('.inner'),
         ]);
 };